Fix warning in XCode

+++ ooniprobe/Utility/LocalizationUtility.m
@@ -96,7 +96,7 @@
     //TODO convert seconds to minutes and hours when needed
     //if getRuntime is MAX_RUNTIME_DISABLED show one hour
     if (runTime <= [MAX_RUNTIME_DISABLED intValue])
-        runTime = SECONDS_IN_HOUR;
+        runTime = (int)SECONDS_IN_HOUR;
     NSString *time = NSLocalizedFormatString(@"Dashboard.Card.Seconds", [NSString stringWithFormat:@"%d", runTime]);
     return time;
 }
